// src/types/inquiriesType.ts
export type Json = string | number | boolean | null | { [key: string]: Json | undefined } | Json[];

export type inquirie = Database['public']['Tables']['user_inquiries']['Row'];
export type inquirieInsert = Database['public']['Tables']['user_inquiries']['Insert'];
export type inquirieUpdate = Database['public']['Tables']['user_inquiries']['Update'];

export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: '13.0.5';
  };
  public: {
    Tables: {
// ...
      user_inquiries: {
        Row: {
          inquiry_answer: string | null;
          inquiry_answered_at: string | null;
          inquiry_created_at: string;
          inquiry_detail: string;
          inquiry_file_urls: Json | null;
          inquiry_id: string;
          inquiry_main_type: string;
          inquiry_status: Database['public']['Enums']['inquiry_status_enum'];
          inquiry_sub_type: string;
          user_id: string;
        };
        Insert: {
          inquiry_answer?: string | null;
          inquiry_answered_at?: string | null;
          inquiry_created_at?: string;
          inquiry_detail: string;
          inquiry_file_urls?: Json | null;
          inquiry_id?: string;
          inquiry_main_type: string;
          inquiry_status?: Database['public']['Enums']['inquiry_status_enum'];
          inquiry_sub_type: string;
          user_id: string;
        };
        Update: {
          inquiry_answer?: string | null;
          inquiry_answered_at?: string | null;
          inquiry_created_at?: string;
          inquiry_detail?: string;
          inquiry_file_urls?: Json | null;
          inquiry_id?: string;
          inquiry_main_type?: string;
          inquiry_status?: Database['public']['Enums']['inquiry_status_enum'];
          inquiry_sub_type?: string;
          user_id?: string;
        };
        Relationships: [
          {
            foreignKeyName: 'user_inquiries_user_id_fkey';
            columns: ['user_id'];
            isOneToOne: false;
            referencedRelation: 'user_profiles';
            referencedColumns: ['user_id'];
          },
        ];
      };
// ...
      user_profiles: {
        Row: {
          agree_email: boolean | null;
          agree_push: boolean | null;
          agree_sms: boolean | null;
          agree_updated_at: string;
          avatar_url: string | null;
          created_at: string;
          email: string;
          is_active: boolean;
          name: string;
          nickname: string;
          user_id: string;
        };
        Insert: {
          agree_email?: boolean | null;
          agree_push?: boolean | null;
          agree_sms?: boolean | null;
          agree_updated_at?: string;
          avatar_url?: string | null;
          created_at?: string;
          email: string;
          is_active?: boolean;
          name: string;
          nickname: string;
          user_id: string;
        };
        Update: {
          agree_email?: boolean | null;
          agree_push?: boolean | null;
          agree_sms?: boolean | null;
          agree_updated_at?: string;
          avatar_url?: string | null;
          created_at?: string;
          email?: string;
          is_active?: boolean;
          name?: string;
          nickname?: string;
          user_id?: string;
        };
        Relationships: [];
      };
    };
// ...
  };
};
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema['Enums']
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals;
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions['schema']]['Enums']
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals;
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions['schema']]['Enums'][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema['Enums']
    ? DefaultSchema['Enums'][DefaultSchemaEnumNameOrOptions]
    : never;